separate files and delete some of tests
