+++ 07.bowling_object/bowling_oop_frame.rb
@@ -1,0 +1,13 @@
+# !/usr/bin/env ruby
+# frozen_string_literal: true
+
+require './bowling_oop_shot'
+
+class Frame
+  attr_reader :first_shot, :second_shot
+
+  def initialize(first_shot, second_shot)
+    @first_shot = Shot.new(first_shot)
+    @second_shot = Shot.new(second_shot)
+  end
+end
