+++ 07.bowling_object/bowling_oop_shot.rb
@@ -1,0 +1,16 @@
+# !/usr/bin/env ruby
+# frozen_string_literal: true
+
+class Shot
+  attr_reader :mark
+
+  def initialize(mark)
+    @mark = mark
+  end
+
+  def score
+    return 10 if mark == 'X'
+
+    mark.to_i
+  end
+end
